校订XY轴电机到达目标所需时间
重新设定通信协议，目标坐标值为16位，设定了一个传输结束的停止控制指令；根据丝杆有效行程，重新计算了到达目标所需时间；重新设定了串口缓冲区结构体变量参数；修改了执行XY轴单步操作以及取印章的判断条件。

+++ HARDWARE/MOTOR_CTRL/motor.h
@@ -1,3 +1,24 @@
+/*
+ *
+ *  Copyright 2018，程豪琪，哈尔滨工业大学
+ *  All Rights Reserved.
+ */
+/*
+ * *******************************郑重声明****************************************
+ 
+ * 该程序为本人所写，付出了大量的精力，现将其公开出来供大家参考学习；
+ * 任何个人和组织不得未经授权将此程序转载，或用于商业行为！
+ * 
+ * 由于本人水平有限，程序难免出现错误，可以通过下面的联系方式联系本人
+ * 谢谢你的指正！
+ *
+ * 邮箱：18s153717@stu.hit.edu.cn
+ * github：https://github.com/clearcumt
+ * 博客：https://www.cnblogs.com/loveclear/
+ * 
+ * ******************************************************************************
+ */
+
 #ifndef __MOTOR_H
 #define __MOTOR_H
 
@@ -17,8 +38,8 @@ typedef __packed struct {
 	/* 外部可变参数 */
 	unsigned char direction;		// 电机方向
 	uint16_t speed;					// 电机速度，改变psc,进而pwm脉冲频率，最后改变速度
-	float destn;					// 电机最终移动目的地坐标
-	float totalTime;				// 电机移动到最终目的地总共所需时间
+	float destn;					// 电机最终移动目的地坐标（mm）
+	int totalTime;					// 电机移动到最终目的地总共所需时间(ms)
 	unsigned char nextMove;
